+++ extensions/spiceblow-database/src/search-database.tsx
@@ -1,3 +1,6 @@
+import fs from "fs";
+import kebabCase from "kebab-case";
+
 import {
   Action,
   ActionPanel,
@@ -14,7 +17,8 @@ import {
   Toast,
   useNavigation,
 } from "@raycast/api";
-import { FormValidation, useForm, usePromise, withAccessToken } from "@raycast/utils";
+
+import { createDeeplink, DeeplinkType, FormValidation, useForm, usePromise, withAccessToken } from "@raycast/utils";
 import { fetch } from "undici";
 
 import { getAvatarIcon, OAuthService, useCachedPromise } from "@raycast/utils";
@@ -45,10 +49,20 @@ import React from "react";
 import { SQLStatement } from "sql-template-strings";
 import { renderColumnValue } from "./database.js";
 import { useGlobalState } from "./state";
-import { CustomQuery, CustomQueryType, GraphData, Json, StoredDatabase, TableInfo, TimeSeriesItem } from "./types";
+import {
+  CustomQuery,
+  CustomQueryType,
+  GraphData,
+  Json,
+  launchContext,
+  StoredDatabase,
+  TableInfo,
+  TimeSeriesItem,
+  type LaunchContext,
+} from "./types";
 
 const pageSize = 20;
-const freeRequestsCount = 10;
+const freeRequestsCount = 50;
 
 let abortController = new AbortController();
 
@@ -114,20 +128,23 @@ function BuyLicense() {
 
 const searchFilterCache = new Cache();
 
-function useSearchFilter(id: string) {
+function useSearchFilter(id: string, defaultOne = "") {
   const [searchText, setSearchText] = useState(() => {
-    return searchFilterCache.get(id) || "";
+    return defaultOne || searchFilterCache.get(id) || "";
   });
   useEffect(() => {
-    searchFilterCache.set(id, searchText);
-  }, [searchText, id]);
+    if (searchText !== defaultOne) {
+      searchFilterCache.set(id, searchText);
+    }
+  }, [searchText, id, defaultOne]);
   return { searchText, setSearchText };
 }
 
 function SearchTable({ table }: { table: string }) {
-  const { searchText, setSearchText } = useSearchFilter(table);
+  const { searchText, setSearchText } = useSearchFilter(table, launchContext?.searchText);
   const currentConnectionString = useGlobalState((x) => x.connectionString);
-  const data = useCachedPromise(
+
+  const tableInfoData = useCachedPromise(
     async (table, _connectionString) => {
       const [schema, tableInfo] = await Promise.all([
         databaseFunctions.getDatabaseSchema({ schemas: [table], tables: [table] }),
@@ -140,9 +157,9 @@ function SearchTable({ table }: { table: string }) {
     {},
   );
 
-  const tableInfo = data.data?.tableInfo;
-  const schema = data.data?.schema;
-
+  const tableInfo = tableInfoData.data?.tableInfo;
+  const schema = tableInfoData.data?.schema;
+  const [isLoading, setIsLoading] = useState(false);
   const { searchField, dropdown } = useTableFiltering({ tableInfo: tableInfo });
 
   const rows = useCachedPromise(
@@ -172,14 +189,54 @@ function SearchTable({ table }: { table: string }) {
     [currentConnectionString, searchText, table],
     { keepPreviousData: true },
   );
-
   const primaryKeyColumns = tableInfo?.columns.filter((col) => col.isPrimaryKey).map((col) => col.columnName) || [];
+  const currentRowIds = new Set(
+    rows.data?.map((row) => primaryKeyColumns.map((col) => row[col]).join(", ")).filter(Boolean) || [],
+  );
+  const selectedRows = useGlobalState((x) =>
+    Array.isArray(x.selectedRows) ? x.selectedRows.filter((row) => currentRowIds.has(row)) : [],
+  );
+
+  const toggleRowSelection = useGlobalState((x) => x.toggleRowSelection);
+  const { push } = useNavigation();
+
+  const handleDeleteSelectedRows = async () => {
+    const deleteQueries = (
+      await Promise.all(
+        selectedRows.map(async (primaryKeyValue) => {
+          const row = rows.data?.find((r) => {
+            return primaryKeyColumns.map((col) => r[col]).join(", ") === primaryKeyValue;
+          });
+          if (!row) {
+            throw new Error(`Could not find row with primary key ${primaryKeyValue}`);
+          }
+          const result = await databaseFunctions.prepareTableRowDelete({
+            allValues: tableInfo!.columns.map((col) => {
+              return { ...col, oldValue: row[col.columnName] };
+            }),
+          });
+          return result.deleteQueries;
+        }),
+      )
+    ).flat();
+
+    push(
+      <DeleteRows
+        revalidate={() => {
+          rows.revalidate();
+        }}
+        deleteQueries={deleteQueries}
+      />,
+    );
+  };
+
+  const [, tableName] = table.split(".");
 
   return (
     <List
       pagination={{ hasMore: false, onLoadMore() {}, ...rows.pagination, pageSize }}
       searchText={searchText}
-      isLoading={rows.isLoading || data.isLoading}
+      isLoading={isLoading || rows.isLoading || tableInfoData.isLoading}
       onSearchTextChange={setSearchText}
       isShowingDetail
       searchBarAccessory={dropdown}
@@ -193,17 +250,53 @@ function SearchTable({ table }: { table: string }) {
     >
       {rows.data?.map((row, index) => {
         const primaryKeyValue = primaryKeyColumns.map((col) => row[col]).join(", ") || String(index);
+        const isSelected = selectedRows?.includes(primaryKeyValue);
+
         return (
           <List.Item
             key={primaryKeyValue}
             title={primaryKeyValue}
-            detail={<List.Item.Detail metadata={tableInfo && <RowMetadata tableInfo={tableInfo} row={row} />} />}
+            icon={selectedRows.length > 0 ? { source: isSelected ? Icon.CheckCircle : Icon.Circle } : undefined}
+            detail={
+              <List.Item.Detail
+                metadata={tableInfo && <RowMetadata tableName={tableName} tableInfo={tableInfo} row={row} />}
+              />
+            }
             actions={
               <ActionPanel>
+                {selectedRows.length > 0 && (
+                  <>
+                    <Action
+                      title={isSelected ? "Deselect Row" : "Select Row"}
+                      icon={isSelected ? Icon.CheckCircle : Icon.Circle}
+                      onAction={() => toggleRowSelection(primaryKeyValue)}
+                    />
+                    <Action
+                      title={`Delete ${selectedRows.length} Selected Row${selectedRows.length > 1 ? "s" : ""}`}
+                      icon={Icon.Trash}
+                      onAction={handleDeleteSelectedRows}
+                      style={Action.Style.Destructive}
+                    />
+                    <Action
+                      title="Deselect All Rows"
+                      icon={Icon.XMarkCircle}
+                      onAction={() => useGlobalState.getState().setSelectedRows([])}
+                    />
+                  </>
+                )}
+
                 {tableInfo && (
                   <RowUpdatesActions revalidate={() => rows.revalidate()} tableInfo={tableInfo} row={row} />
                 )}
+                {!selectedRows.length && (
+                  <Action
+                    title={"Select Row"}
+                    icon={Icon.CheckCircle}
+                    onAction={() => toggleRowSelection(primaryKeyValue)}
+                  />
+                )}
                 <CountAction table={table} />
+                <ExportToCsvAction tableInfo={tableInfo!} setIsLoading={setIsLoading} table={table} />
                 <RunTransactionQueries />
               </ActionPanel>
             }
@@ -225,6 +318,85 @@ function CountAction({ table, query }: { table?: string; query?: string }) {
   };
 
   return <Action title="Count Rows" icon={Icon.List} onAction={handleCount} />;
+}
+
+function ExportToCsvAction({
+  table = "",
+  setIsLoading,
+  query = "",
+  tableInfo,
+}: {
+  table?: string;
+  setIsLoading: (loading: boolean) => void;
+  query?: string;
+  tableInfo: TableInfo;
+}) {
+  const handleExport = async () => {
+    setIsLoading(true);
+    try {
+      const databases = JSON.parse((await LocalStorage.getItem("databases")) || "[]");
+      const currentConnectionString = useGlobalState.getState().connectionString;
+      const currentDb = databases.find((db) => db.connectionString === currentConnectionString);
+      const dbName = currentDb?.name || "";
+      const tempFile = `/tmp/export-${dbName ? `${kebabCase(dbName)}-` : ""}${table ? `${kebabCase(table)}-` : ""}${Date.now()}.csv`;
+
+      let page = 0;
+      let hasMore = true;
+      const pageSize = 3000;
+      let isFirstPage = true;
+      let totalRows = 0;
+      while (hasMore) {
+        console.log(`fetching page ${page}`);
+        showToast({
+          style: Toast.Style.Animated,
+          title: `Fetching page ${page + 1}...`,
+        });
+        const { data: rows, hasMore: moreRows } = await databaseFunctions.searchTableRowsOrCustomQuery({
+          table,
+          query,
+          page,
+          pageSize,
+        });
+
+        totalRows += rows.length;
+
+        // Convert all values to CSV-friendly strings
+        for (const row of rows) {
+          for (const column of tableInfo?.columns || []) {
+            row[column.columnName] = databaseFunctions.renderColumnValueForCsv(column, row[column.columnName]);
+          }
+        }
+        // Generate CSV for this batch, only include headers on first page
+        const csvChunk = databaseFunctions.exportToCsv({
+          rows,
+          includeHeader: isFirstPage,
+        });
+
+        // Append to file (or write if first page)
+        await fs.promises.writeFile(tempFile, csvChunk + "\n", { flag: isFirstPage ? "w" : "a" });
+
+        hasMore = moreRows;
+        page++;
+        isFirstPage = false;
+      }
+      console.log(`fetched ${totalRows} rows`);
+
+      const file = tempFile;
+      await Clipboard.copy({ file });
+      showToast({ style: Toast.Style.Success, title: "CSV file copied to clipboard" });
+    } catch (error) {
+      console.error("Error exporting to CSV:", error);
+      showToast({ style: Toast.Style.Failure, title: "Failed to export CSV", message: String(error) });
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <>
+      <Action title="Export as Csv File" icon={Icon.Download} onAction={() => handleExport()} />
+    </>
+  );
 }
 
 function NoRowsActions({ revalidate, tableInfo }: { revalidate: () => void; tableInfo: TableInfo }) {
@@ -277,7 +449,7 @@ function RowUpdatesActions({
             }),
           });
 
-          push(<DeleteRow revalidate={revalidate} deleteQueries={deleteQueries} />);
+          push(<DeleteRows revalidate={revalidate} deleteQueries={deleteQueries} />);
         }}
       />
 
@@ -404,6 +576,9 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
   const { searchText, setSearchText } = useSearchFilter(id);
 
   const currentConnectionString = useGlobalState((x) => x.connectionString);
+
+  const toggleRowSelection = useGlobalState((x) => x.toggleRowSelection);
+
   const tables = [...new Set(tableInfo.columns.map((col) => col.tableName).filter(isTruthy))];
   const schema = useCachedPromise(
     async (_connectionString, selectedSchemas) => {
@@ -442,6 +617,45 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
     [currentConnectionString, searchText, query, searchField, tableInfo, schema.data],
     { keepPreviousData: true },
   );
+  const currentRowIds = new Set(rows.data?.map((row) => String(row[bestField])).filter(Boolean) || []);
+  const selectedRows = useGlobalState((x) =>
+    Array.isArray(x.selectedRows) ? x.selectedRows.filter((row) => currentRowIds.has(row)) : [],
+  );
+
+  const [isLoading, setIsLoading] = useState(false);
+  const handleDeleteSelectedRows = async (tableName?: string) => {
+    const deleteQueries = (
+      await Promise.all(
+        selectedRows.map(async (primaryKeyValue) => {
+          const row = rows.data?.find((r) => {
+            return String(r[bestField]) === primaryKeyValue;
+          });
+          if (!row) {
+            throw new Error(
+              `Could not find row with value ${primaryKeyValue} in field ${bestField}. Full row data: ${JSON.stringify(rows.data)}`,
+            );
+          }
+          const result = await databaseFunctions.prepareTableRowDelete({
+            allValues: tableInfo.columns
+              .filter((col) => !tableName || col.tableName === tableName)
+              .map((col) => {
+                return { ...col, oldValue: row[col.columnName] };
+              }),
+          });
+          return result.deleteQueries;
+        }),
+      )
+    ).flat();
+
+    push(
+      <DeleteRows
+        revalidate={() => {
+          rows.revalidate();
+        }}
+        deleteQueries={deleteQueries}
+      />,
+    );
+  };
 
   if (args.type === "time-series") {
     return <Graph query={args} rows={rows.data} />;
@@ -450,7 +664,7 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
   return (
     <List
       pagination={{ hasMore: false, onLoadMore() {}, ...rows.pagination, pageSize }}
-      isLoading={rows.isLoading}
+      isLoading={isLoading || rows.isLoading}
       searchText={searchText}
       onSearchTextChange={setSearchText}
       isShowingDetail
@@ -465,18 +679,50 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
     >
       {rows.data?.map((row, index) => {
         const primaryKeyValue = String(row[bestField]) || String(index);
+        const isSelected = selectedRows?.includes(primaryKeyValue);
+
         return (
           <List.Item
             key={primaryKeyValue + index}
-            title={primaryKeyValue || " "} // JSON stringified object in the list title
+            title={primaryKeyValue || " "}
+            icon={selectedRows.length > 0 ? { source: isSelected ? Icon.CheckCircle : Icon.Circle } : undefined}
             detail={
               <List.Item.Detail
-                // markdown={" "}
-                metadata={tableInfo && <RowMetadata tableInfo={tableInfo} row={row} />}
+                metadata={tableInfo && <RowMetadata tableName={"query"} tableInfo={tableInfo} row={row} />}
               />
             }
             actions={
               <ActionPanel>
+                {selectedRows.length > 0 && (
+                  <>
+                    <Action
+                      title={isSelected ? "Deselect Row" : "Select Row"}
+                      icon={isSelected ? Icon.CheckCircle : Icon.Circle}
+                      onAction={() => toggleRowSelection(primaryKeyValue)}
+                    />
+                    {tables.map((tableName) => (
+                      <Action
+                        key={tableName}
+                        title={`Delete ${selectedRows.length} in "${tableName}"`}
+                        icon={Icon.Trash}
+                        onAction={() => handleDeleteSelectedRows(tableName)}
+                        style={Action.Style.Destructive}
+                      />
+                    ))}
+                    <Action
+                      title={`Delete ${selectedRows.length} Row${selectedRows.length > 1 ? "s" : ""} in All Tables`}
+                      icon={Icon.Trash}
+                      onAction={() => handleDeleteSelectedRows()}
+                      style={Action.Style.Destructive}
+                    />
+                    <Action
+                      title="Deselect All Rows"
+                      icon={Icon.XMarkCircle}
+                      onAction={() => useGlobalState.getState().setSelectedRows([])}
+                    />
+                  </>
+                )}
+
                 <RowUpdatesActions
                   revalidate={() => {
                     rows.revalidate();
@@ -484,7 +730,17 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
                   tableInfo={tableInfo}
                   row={row}
                 />
+
+                {!selectedRows.length && (
+                  <Action
+                    title={"Select Row"}
+                    icon={Icon.CheckCircle}
+                    onAction={() => toggleRowSelection(primaryKeyValue)}
+                  />
+                )}
+
                 <CountAction query={query} />
+                <ExportToCsvAction tableInfo={tableInfo} setIsLoading={setIsLoading} query={query} />
                 {tables.map((tableName) => (
                   <Action
                     key={tableName}
@@ -500,7 +756,7 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
                       });
 
                       push(
-                        <DeleteRow
+                        <DeleteRows
                           revalidate={() => {
                             rows.revalidate();
                           }}
@@ -537,8 +793,7 @@ function SearchCustomQuery(args: CustomQuery & { revalidate: () => void }) {
     </List>
   );
 }
-
-function RowMetadata({ tableInfo, row }: { tableInfo: TableInfo; row: Json }) {
+function RowMetadata({ tableName, tableInfo, row }: { tableName: string; tableInfo: TableInfo; row: Json }) {
   return (
     <List.Item.Detail.Metadata>
       {tableInfo?.columns?.map((col, idx) => {
@@ -575,6 +830,35 @@ function RowMetadata({ tableInfo, row }: { tableInfo: TableInfo; row: Json }) {
           <Fragment key={idx}>
             {item}
             {!isLast && <List.Item.Detail.Metadata.Separator key={`${idx}-separator`} />}
+          </Fragment>
+        );
+      })}
+
+      {tableInfo?.relations?.map((relation, idx) => {
+        return (
+          <Fragment key={`relation-${idx}`}>
+            <List.Item.Detail.Metadata.Separator />
+            <List.Item.Detail.Metadata.Link
+              title={`${relation.foreignTable}`}
+              text={relation.columnNames
+                .map((col, i) => `${tableName}.${col} → ${relation.foreignTable}.${relation.foreignColumnNames[i]}`)
+                .join(", ")}
+              target={createDeeplink({
+                extensionName: "spiceblow-database",
+                context: {
+                  searchText: relation.foreignColumnNames
+                    .map((field, i) => {
+                      const value = String(row[relation.columnNames[i]]);
+                      return `${field} is ${value}`;
+                    })
+                    .join(" and "),
+                  schemaTable: `${relation.foreignSchema}.${relation.foreignTable}`,
+                  view: "search-table",
+                } satisfies LaunchContext,
+                type: DeeplinkType.Extension,
+                command: "search-database",
+              })}
+            />
           </Fragment>
         );
       })}
@@ -1071,7 +1355,7 @@ function SearchTables() {
         return null;
       }
       const [schema, tableName] = table.schemaTable.split(".");
-      const schemaColor = getStringColor(schema || "", { saturation: 0.2 });
+      const schemaColor = getStringColor("_" + schema || "", { saturation: 0.4 });
       const tableColor = getStringColor(table.schemaTable);
 
       return {
@@ -1080,12 +1364,12 @@ function SearchTables() {
           <List.Item
             key={table.schemaTable}
             title={tableName}
-            subtitle={`about ${table.estimatedRowCount} rows`}
+            // subtitle={table.columns?.map((c) => c).join(", ") || ""}
             icon={{
               source: Icon.Document,
               tintColor: tableColor,
             }}
-            keywords={[schema]}
+            keywords={[schema, ...(table.columns || [])]}
             accessories={[
               {
                 tag: { value: schema, color: schemaColor },
@@ -1275,19 +1559,19 @@ function useDatabases() {
   return databases;
 }
 
-function DeleteRow({ revalidate, deleteQueries }: { revalidate: () => void; deleteQueries: SQLStatement[] }) {
+function DeleteRows({ revalidate, deleteQueries }: { revalidate: () => void; deleteQueries: SQLStatement[] }) {
   const { pop } = useNavigation();
   const [isLoading, setIsLoading] = useState(false);
   const { addQueriesToTransaction } = useGlobalState();
   const queriesMarkdown = getQueryMarkdown(deleteQueries);
-  const markdown = "Are you sure you want to delete this row?\n\n" + queriesMarkdown;
+  const markdown = "Are you sure you want to execute the following queries?\n\n" + queriesMarkdown;
 
   async function executeDeleteQuery() {
     try {
       setIsLoading(true);
       await databaseFunctions.executeQueries({ queries: deleteQueries });
       await revalidate();
-      showToast({ title: "Row deleted successfully", style: Toast.Style.Success });
+      showToast({ title: "Deleted successfully", style: Toast.Style.Success });
       pop();
     } catch (error) {
       notifyError(error);
@@ -1299,11 +1583,11 @@ function DeleteRow({ revalidate, deleteQueries }: { revalidate: () => void; dele
   return (
     <Detail
       isLoading={isLoading}
-      navigationTitle="Delete Row?"
+      navigationTitle="Delete Rows?"
       markdown={markdown}
       actions={
         <ActionPanel>
-          <Action title="Delete Row" onAction={executeDeleteQuery} icon={Icon.Trash} style={Action.Style.Destructive} />
+          <Action title="Delete" onAction={executeDeleteQuery} icon={Icon.Trash} style={Action.Style.Destructive} />
           <Action
             title="Add to Transaction"
             onAction={() => {
@@ -1478,10 +1762,11 @@ function EditRow({
         const canBeUpdated = column.originalColumnName && column.tableName;
         const value = renderColumnValue(column, formValues[column.columnName]);
 
-        let C = Form.TextField;
+        let FormComponent = Form.TextField;
         if (value.split("\n").length > 1) {
-          C = Form.TextArea;
+          FormComponent = Form.TextArea;
         }
+
         let error = "";
         if (requiredFields.includes(column)) {
           if (!value) {
@@ -1501,14 +1786,34 @@ function EditRow({
         if (!requiredFields.includes(column)) {
           const col = tableInfo.columns.find((x) => x.columnName === column.columnName);
           if (col?.defaultValue) {
-            placeholder = `${col?.defaultValue}`;
+            placeholder = col?.defaultValue;
           } else {
             placeholder = `null`;
           }
         }
 
+        if (column.enumValues?.length) {
+          // console.log('column.enumValues', column.enumValues);
+          return (
+            <Form.Dropdown
+              key={column.columnName}
+              id={column.columnName}
+              error={error}
+              title={column.columnName}
+              info={info}
+              defaultValue={value || ""}
+              onChange={(newValue) => handleFieldChange(column.columnName, newValue)}
+            >
+              {column.isNullable && <Form.Dropdown.Item key="null" value="" title="null" />}
+              {column.enumValues?.map((enumValue) => (
+                <Form.Dropdown.Item key={enumValue} value={enumValue} title={enumValue} />
+              ))}
+            </Form.Dropdown>
+          );
+        }
+
         return (
-          <C
+          <FormComponent
             key={column.columnName}
             id={column.columnName}
             error={error}
@@ -1564,6 +1869,7 @@ function DatabasesDropdown() {
           );
           return;
         }
+
         setConnectionString(newValue);
       }}
     >
@@ -1730,6 +2036,10 @@ const Command = ({ draftValues: draftValuesArg }: LaunchProps<{ draftValues?: Ne
 
   if (!connectionString) {
     return <List isLoading={true} />;
+  }
+
+  if (launchContext?.view === "search-table") {
+    return <SearchTable table={launchContext.schemaTable} />;
   }
 
   return <SearchTables key={connectionString} />;