+++ extensions/toggl-track/src/index.tsx
@@ -8,6 +8,7 @@ import RunningTimeEntry from "@/components/RunningTimeEntry";
 import UpdateTimeEntryForm from "@/components/UpdateTimeEntryForm";
 import { ExtensionContextProvider } from "@/context/ExtensionContext";
 import { formatSeconds } from "@/helpers/formatSeconds";
+import Shortcut from "@/helpers/shortcuts";
 import { Verb, withToast } from "@/helpers/withToast";
 import { useProcessedTimeEntries } from "@/hooks/useProcessedTimeEntries";
 import { useTimeEntryActions } from "@/hooks/useTimeEntryActions";
@@ -73,9 +74,9 @@ function ListView() {
               title={timeEntry.description || "No description"}
               subtitle={(timeEntry.client_name ? timeEntry.client_name + " | " : "") + (timeEntry.project_name ?? "")}
               accessories={[
-                { text: formatSeconds(timeEntry.duration) },
                 ...timeEntry.tags.map((tag) => ({ tag })),
                 timeEntry.billable ? { tag: { value: "$" } } : {},
+                { text: formatSeconds(timeEntry.duration) },
               ]}
               icon={{ source: Icon.Circle, tintColor: timeEntry.project_color }}
               actions={
@@ -97,6 +98,7 @@ function ListView() {
                   <Action.Push
                     title="Create Similar Time Entry"
                     icon={{ source: Icon.Plus }}
+                    shortcut={Shortcut.Duplicate}
                     target={
                       <ExtensionContextProvider>
                         <TimeEntryForm