+++ extensions/twenty/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Twenty Changelog
 
+## [Maintenance] - 2025-02-12
+
+- Cleanup unused `.github` files
+
+## [Enhancements] - 2025-01-09
+
+- Added support for multi-select field type
+
+## [Enhancements] - 2024-12-18
+
+- Updated select field type with preference
+
+## [Enhancements] - 2024-12-11
+
+- Added support for rating field type
+
 ## [Enhancements] - 2024-12-04
 
 - Updated entire twenty extension to be flexible